fix: allow for multiple sessions per identity, allows decrypting older messages
